requirements/dev.txt
Django==4.2
django-crispy-forms
djangorestframework